feat: batch file transfer
Implement batch file transfer, and make some necessary cli changes for
integration test:

1. `--port` (default: `/dev/ttyS0`)
2. Files are declared positional parameters and there can be many of
   them.

In order to make sure that this all is actually working mplement a test
that simulates batch file transfer on a serial line and BPS throttled to
realistic measures. This requires also updating szm to accept multiple
files.

This will require some supplemental fixes to fine-tune some glitches
but only for the not-before-existing multi file scenario.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>